relocate key module to pkg
